added routes to start/stop reminder to index.js
