+++ app.js
@@ -6,7 +6,6 @@ var logger = require("morgan");
 var indexRouter = require("./routes/index");
 var usersRouter = require("./routes/users");
 var apiRouter = require("./routes/api");
-
 var app = express();
 
 app.use(logger("dev"));
@@ -18,5 +17,6 @@ app.use(express.static(path.join(__dirname, "public")));
 app.use("/", indexRouter);
 app.use("/users", usersRouter);
 app.use("/api", apiRouter);
+  
 
 module.exports = app;
