+++ package.json
@@ -2,7 +2,7 @@
   "name": "my-express-app.",
   "version": "0.0.0",
   "private": true,
-  "type": "module",
+
   "scripts": {
     "start": "node ./bin/www",
     "migrate": "node model/database.js"
